Split implementations from the interfaces for Artifact.
