+++ source/slang/slang.cpp
@@ -7,7 +7,8 @@
 #include "../core/slang-type-text-util.h"
 #include "../core/slang-type-convert-util.h"
 
-#include "../compiler-core/slang-artifact.h"
+// Artifact
+#include "../compiler-core/slang-artifact-impl.h"
 #include "../compiler-core/slang-artifact-desc-util.h"
 
 #include "slang-module-library.h"
